+++ R/tokens.R
@@ -11,7 +11,7 @@
 #' not be possible to remove things that are not present.  For instance, if the
 #' `tokens` object has already had punctuation removed, then `tokens(x,
 #' remove_punct = TRUE)` will have no additional effect.
-#' @param x the input object to the tokens constructor, one of: 
+#' @param x the input object to the tokens constructor, one of:
 #'   a (uniquely) named **list** of characters; a [tokens] object; or a
 #'   [corpus] or [character] object that will be tokenized
 #' @param what character; which tokenizer to use.  The default `what = "word"`
@@ -71,13 +71,13 @@
 #'   for hashtags and
 #'   [here](https://help.twitter.com/en/managing-your-account/twitter-username-rules)
 #'   for usernames.
-#'   
+#'
 #'   In versions < 2, the argument `remove_twitter` controlled whether social
 #'   media tags were preserved or removed, even when `remove_punct = TRUE`.
 #'   This argument is not longer functional in versions >= 2.  If greater
 #'   control over social media tags is desired, you should user an alternative
 #'   tokenizer, including non-\pkg{quanteda} options.
-#'   
+#'
 #'   For backward compatibility, the following older tokenizers are also
 #'   supported through `what`:
 #'   \describe{
@@ -242,7 +242,7 @@ tokens.corpus <- function(x,
                               "fasterword", "fastestword"))
     # deprecated arguments
     if ("remove_hyphens" %in% names(dots)) {
-        split_hyphens <- dots$remove_hyphens
+        split_hyphens <- dots[["remove_hyphens"]]
         .Deprecated(msg = "'remove_hyphens' is deprecated, use 'split_hyphens' instead.")
         dots$remove_hyphens <- NULL
     }
@@ -262,44 +262,51 @@ tokens.corpus <- function(x,
                            character = tokenize_character,
                            fasterword = tokenize_fasterword,
                            fastestword = tokenize_fastestword)
-    
-    
-    if (what == "word") {
-        x <- preserve_special(x, split_hyphens = split_hyphens, 
-                               split_tags = FALSE, verbose = verbose)
-        special <- attr(x, "special")
-    }
-    
-    # split x into smaller blocks to reducre peak memory consumption
-    x <- split(x, ceiling(seq_along(x) / 10000))
-    for (i in seq_along(x)) {
-        if (verbose) catm(" ...tokenizing", i, "of", length(x), "blocks\n")
-        temp <- tokenizer_fn(x[[i]], split_hyphens = split_hyphens, verbose = verbose)
-        if (verbose) catm(" ...serializing tokens, ")
-        if (i == 1) {
-            x[[i]] <- serialize_tokens(temp)
-        } else {
-            x[[i]] <- serialize_tokens(temp, attr(x[[i - 1]], "types"))
+
+
+    if (!remove_separators && !what %in% c("word", "word1", "character"))
+        warning("remove_separators is always TRUE for this type")
+
+    # split x into smaller blocks to reduce peak memory consumption
+    x <- texts(x)
+    x <- split(x, factor(ceiling(seq_along(x) / quanteda_options("tokens_block_size"))))
+    x <- lapply(x, function(y) {
+        if (verbose)
+            catm(" ...", head(names(y), 1), " to ", tail(names(y), 1), "\n", sep = "")
+            #catm(" ...", head(names(y), 1), " to ", tail(names(y), 1),
+            #     " by process ", Sys.getpid(), "\n", sep = "")
+            
+        y <- normalize_characters(y)
+        if (what == "word") {
+            y <- preserve_special(y, split_hyphens = split_hyphens,
+                                  split_tags = FALSE, verbose = verbose)
+            special <- attr(y, "special")
         }
-        if (verbose) catm(format(length(attr(x[[i]], "types")), big.mark = ",", trim = TRUE),
-                          "unique types\n")
-    }
-    
+        y <- serialize_tokens(tokenizer_fn(y, split_hyphens = split_hyphens, verbose = verbose))
+        if (what == "word")
+            y <- restore_special(y, special)
+        if (what == "word1" && !split_hyphens)
+            y <- restore_special1(y, split_hyphens = FALSE, split_tags = TRUE)
+        return(y)
+    })
+    type <- unique(unlist(lapply(x, attr, "types"), use.names = FALSE))
+    if (verbose)
+        catm(" ...", format(length(type), big.mark = ",", trim = TRUE),
+             " unique type", if (length(type) == 1) "" else "s",
+             "\n", sep = "")
+    x <- lapply(x, function(y) {
+        map <- fastmatch::fmatch(attr(y, "types"), type)
+        y <- lapply(y, function(z) map[z])
+        return(y)
+    })
+
     result <- build_tokens(
-        unlist(x, recursive = FALSE), 
-        types = attr(x[[length(x)]], "types"),
-        what = what, 
+        unlist(x, recursive = FALSE),
+        types = type, what = what,
         docvars = select_docvars(attrs[["docvars"]], user = include_docvars, system = TRUE),
         meta = attrs[["meta"]]
     )
 
-    if (!remove_separators && !what %in% c("word", "word1", "character"))
-        warning("remove_separators is always TRUE for this type")
-    if (what == "word")
-        result <- restore_special(result, special)
-    if (what == "word1" && !split_hyphens)
-        result <- restore_special1(result, split_hyphens = FALSE, split_tags = TRUE)
-    
     result <- tokens.tokens(result,
                             remove_punct = remove_punct,
                             remove_symbols = remove_symbols,
@@ -331,17 +338,17 @@ tokens.tokens <-  function(x,
                            ...) {
     x <- as.tokens(x)
     dots <- list(...)
-    
+
     # deprecated arguments
     if ("remove_hyphens" %in% names(dots)) {
-        split_hyphens <- dots$remove_hyphens
+        split_hyphens <- dots[["remove_hyphens"]]
         .Deprecated(msg = "'remove_hyphens' is deprecated, use 'split_hyphens' instead.")
         dots$remove_hyphens <- NULL
 
     }
     if ("remove_twitter" %in% names(dots)) {
         .Deprecated(msg = "'remove_twitter' is deprecated and inactive for tokens.tokens()")
-        dots$remove_twitter <- NULL    
+        dots$remove_twitter <- NULL
     }
     check_dots(dots, c(names(formals(tokens))))
 
@@ -350,37 +357,39 @@ tokens.tokens <-  function(x,
         if (verbose) catm(" ...splitting hyphens\n")
         x <- tokens_split(x, "\\p{Pd}", valuetype = "regex", remove_separator = FALSE)
     }
-    
+
     # removals
     removals <- removals_regex(separators = remove_separators,
                                punct = remove_punct,
                                symbols = remove_symbols,
                                numbers = remove_numbers,
                                url = remove_url)
-    
+
     if (length(removals) && verbose) {
-        msg <- stri_replace_all_fixed(names(removals), 
-                                      c("url", "punct"), 
+        msg <- stri_replace_all_fixed(names(removals),
+                                      c("url", "punct"),
                                       c("URLs", "punctuation"),
                                       vectorize_all = FALSE)
         catm(" ...removing", paste(msg, collapse = ", "), "\n")
     }
-    
+
     if (length(removals[["separators"]])) {
-        x <- tokens_remove(x, removals[["separators"]], valuetype = "regex")
+        x <- tokens_remove(x, removals[["separators"]], valuetype = "regex",
+                           verbose = FALSE)
         removals["separators"] <- NULL
     }
-    
+
     if (length(removals)) {
         x <- tokens_remove(x, paste(unlist(removals), collapse = "|"),
-                           valuetype = "regex",  padding = padding)
+                           valuetype = "regex",  padding = padding,
+                           verbose = FALSE)
     }
 
     if (!include_docvars)
         docvars(x) <- NULL
 
     if (verbose) {
-        catm(" ...complete, elapsed time: ",
+        catm(" ...complete, elapsed time:",
              format((proc.time() - tokens_env$START_TIME)[3], digits = 3), "seconds.\n")
         catm("Finished constructing tokens from ", format(length(x), big.mark = ","), " document",
              if (length(x) > 1) "s", ".\n", sep = "")
@@ -438,8 +447,8 @@ as.tokens.list <- function(x, concatenator = "_", ...) {
     x <- lapply(x, stri_trans_nfc)
     x <- serialize_tokens(x)
     build_tokens(
-        x, 
-        types = attr(x, "types"), 
+        x,
+        types = attr(x, "types"),
         concatenator = concatenator,
         docvars = make_docvars(length(x), names(x))
     )
@@ -486,15 +495,15 @@ removals_regex <- function(separators = FALSE,
                            url = FALSE) {
     regex <- list()
     if (separators)
-        regex[["separators"]] = "^[\\p{Z}\\p{C}]+$"
+        regex[["separators"]] <- "^[\\p{Z}\\p{C}]+$"
     if (punct)
-        regex[["punct"]] = "^\\p{P}+$"
+        regex[["punct"]] <- "^\\p{P}+$"
     if (symbols)
-        regex[["symbols"]] = "^\\p{S}$"
+        regex[["symbols"]] <- "^\\p{S}$"
     if (numbers) # includes currency amounts and those containing , or . digit separators, and 100bn
-        regex[["numbers"]] = "^\\p{Sc}{0,1}\\p{N}+([.,]*\\p{N})*\\p{Sc}{0,1}$"
+        regex[["numbers"]] <- "^\\p{Sc}{0,1}\\p{N}+([.,]*\\p{N})*\\p{Sc}{0,1}$"
     if (url)
-        regex[["url"]] = "(^((https{0,1}|s{0,1}ftp)://)|(\\w+@\\w+))"
+        regex[["url"]] <- "(^((https{0,1}|s{0,1}ftp)://)|(\\w+@\\w+))"
     return(regex)
 }
 
@@ -580,45 +589,45 @@ tokens_recompile <- function(x, method = c("C++", "R"), gap = TRUE, dup = TRUE) 
 
     method <- match.arg(method)
     attrs <- attributes(x)
-
+    type <- attr(x, "types")
     if (method == "C++") {
-        x <- qatd_cpp_tokens_recompile(x, types(x), gap, dup)
+        x <- qatd_cpp_tokens_recompile(x, type, gap, dup)
         x <- rebuild_tokens(x, attrs)
     } else {
-    
+
         # Check for padding
         index_unique <- unique(unlist(unclass(x), use.names = FALSE))
         padding <- index_unique == 0
         attrs[["padding"]] <- any(padding) # add padding flag
         index_unique <- index_unique[!padding] # exclude padding
-    
+
         if (!gap && !dup) return(x)
-    
+
         # Remove gaps in the type index, if any, remap index
         if (gap) {
-            if (any(is.na(match(seq_len(length(types(x))), index_unique)))) {
-                types_new <- types(x)[index_unique]
+            if (any(is.na(match(seq_len(length(type)), index_unique)))) {
+                type_new <- type[index_unique]
                 index_new <- c(0, seq_along(index_unique)) # padding index is zero but not in types
                 index_unique <- c(0, index_unique) # padding index is zero but not in types
                 x <- lapply(unclass(x), function(y) index_new[fastmatch::fmatch(y, index_unique)])
                 attributes(x) <- attrs
-                types(x) <- types_new
+                type <- type_new
             }
         }
-    
+
         # Reindex duplicates, if any
         if (dup) {
-            if (any(duplicated(types(x)))) {
-                types <- types(x)
-                types_unique <- unique(types)
-                index_mapping <- match(types, types_unique)
+            if (any(duplicated(type))) {
+                type_unique <- unique(type)
+                index_mapping <- match(type, type_unique)
                 index_mapping <- c(0, index_mapping) # padding index is zero but not in types
                 x <- lapply(unclass(x), function(y) index_mapping[y + 1]) # shift index for padding
                 attributes(x) <- attrs
-                types(x) <- types_unique
+                type <- type_unique
             }
         }
-        Encoding(types(x)) <- "UTF-8"
+        Encoding(type) <- "UTF-8"
+        attr(x, "types") <- type
         x <- rebuild_tokens(x, attrs)
     }
     return(x)